[BIOCAT-638]
Implemented functionaliy to allow users to add ONE custom endpoint for monitoring.  The actual endpoint will be kept out of view from the public.  Users can update the endpoint being monitoring (at the expense of monitoring history for that endpoint).  The service test can also be enabled and disabled like the others.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/branches/v1.1.0@2463 1acaf488-1b9d-4964-8077-0a2fdb83ef00